+++ src/index.js
@@ -1,3 +1,7 @@
 import PDFEdit from "./editor.js";
+import PdfViewerApi from "./api.js";
 
-PDFEdit.register();
+exports.activate = function (context) {
+  PDFEdit.register();
+  return PdfViewerApi;
+};